Restrict default seed range, to enable manually recovery of a past seed.
When generating a document with random values for an exam,
it's useful to keep track of seed to generate a document with the answer.

Best way to do that is to manually choose seed, so the same pseudo-random
values will be generated in both documents.

However, it was much too easy to forget to choose seed.
Then, since there are billions of billions of possible seeds, it was
impossible to guess which seed was used.

To prevent this, seed range is now restricted by default to 100 000
different possibilities, which is largely enough for practical purpose,
and still enables to guess used seed in a few seconds on a modern
computer (by running a few lines of python code in interpreter).